microproyecto 2.3/miniProyecto2/img-classifier/requirements.txt
flask==3.0.3
pillow==10.4.0
tensorflow-cpu==2.15.0
